Build search bar, retrieve city object
Here I built a simple search bar and added some CSS to get a feel for how the project should look when finished.

I retrieved the open weather object from the API for Seoul, ready for testing.

I also added an event listener to the search bar so that on focus, the text and Google's default outline disappears.