Add weather data code
I've practiced making this weather application twice during my bootcamp. This is my second try of the code, copied and integrated into GitHub.

The for loop in filterWeatherData was learnt from my camp's lecturer. It took some analysis to really understand that i+=8 means that we are jumping across one day. The weather data has an associative array for 8 times in the day, so increasing the iteration by 8 and mutating the number is a great approach from her.

Integrating weatherArray into the loop was my mentor's idea. i becomes the index for the day to receive data from, and j becomes the index for each associative array within the returned weather object. It's really quite clever.

I broke down his code, wrote a how to do guide and then rebuilt the code for myself. filterWeatherData is the closest to his original design.

The key thing I learned in the second try was to use objects. My mentor gave me his version of the code, and this my attempt to use his ideas and structure (which follow an MVC structure and filter arguments through 6 different functions).

While I found it too difficult to build the setToday/createDayDefault/setOtherDay functions exactly the way of my mentors but I did use many of his ideas, in particular his use of arguments.

My bootcamp lecturer showed us a build she did all the assignment and item creation was built through loops and id assignment, but I found objects an easier concept to grasp, and she encouraged me to try it.

With the DOM elements, it seems key to try and avoid repetitive code as much as possible. createItem and createIcon are great tools that I found very useful for feeding arguments into. By using tags and content as an argument, I can feed in whatever data I need.

With the day functions setToday, setOtherDay, setTodayContainer etc, it was very useful to use the parent as an argument. by adding the parent on call of the function, you can assign the parent repeatedly inside the function.

Overall this was a satisfying thing to build.

For GitHub, I'm focusing on added functionality. I've styled the app in a way I like, but I hope to go further and add the background changing feature from the Star Wars quotes app too.