Implement Keyboard controls
This has been an interesting challenge. To complete it, I borrowed code from W3 schools and broke it down before using it. What was challenging was making sure that the text in the input field was changed, but without accidentally submitting the form. Using e-prevent default helped, but the real success came from adjusting the content of my functions. By dividing empty auto results into empty auto results and empty inner HTML, I removed the emptying from the enter function, so that when the user presses enter, the value remains in the input field, but in other functions that need to empty the field, the functionality is still there.

This will likely be my last update for this project, as it's time to move on to more complicated work. I expect to come back to weather app though, and find new way of handling the problem.